+++ modules/cirrus/builtin-functions/inputs.tf
@@ -216,3 +216,15 @@ variable "deploy_alarms" {
   default     = true
   description = "Deploy Cirrus Alarms stack"
 }
+
+variable "domain_alias" {
+  description = "Custom domain alias for private API Gateway endpoint"
+  type        = string
+  default     = ""
+}
+
+variable "private_certificate_arn" {
+  description = "Private Certificate ARN for custom domain alias of private API Gateway endpoint"
+  type        = string
+  default     = ""
+}