+++ modules/stac-server/inputs.tf
@@ -468,4 +468,15 @@ variable "warning_sns_topic_arn" {
 variable "critical_sns_topic_arn" {
   description = "SNS topic to be used by all stac-server `critical` alarms"
   type        = string
+
+variable "domain_alias" {
+  description = "Custom domain alias for private API Gateway endpoint"
+  type        = string
+  default     = ""
+}
+
+variable "private_certificate_arn" {
+  description = "Private Certificate ARN for custom domain alias of private API Gateway endpoint"
+  type        = string
+  default     = ""
 }