+++ CHANGELOG.md
@@ -15,6 +15,35 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Removed
 
+## [2.47.0] - 2025-05-16
+
+### Added
+
+- Support for PRIVATE-type TiTiler API Gateways
+- Added Custom Domain Name support for TiTiler, Stac and Cirrus Private API Gateways
+
+### Changed
+
+- Enforced character limit on cirrus `resource_prefix`
+
+## [2.46.0] - 2025-05-15
+
+### Added
+
+- Added capability to pass a custom S3 bucket for Console UI and Cirrus Dashboard S3 websites
+
+## [2.45.0] - 2025-05-12
+
+### Added
+
+- Added outputs for Console UI and Cirrus Dashboard S3 Bucket names
+
+## [2.44.0] - 2025-05-07
+
+### Fixed
+
+- stac_server: set ENABLE_INGEST_ACTION_TRUNCATE on the ingest lambda rather than the api lambda
+
 ## [2.43.0] - 2025-04-30
 
 ### Added